remove cpp file, add xc7a50t
