+++ examples/basic/Makefile
@@ -1,4 +1,4 @@
 all: main
 
-main: main.cpp ../../idstring.cpp
-	g++ -o main -I../../ main.cpp ../../idstring.cpp
+main: main.cpp ../../idstring.h
+	g++ -o main -I../../ main.cpp
